Dalamud SDK 15 bump. Also fixed stacking issues.

+++ Loci/Processors/Memory/Memory.Gearset.cs
@@ -10,20 +10,20 @@ namespace Loci.Processors;
 // Gearset handling
 public unsafe partial class LociMemory
 {
-    public delegate nint GearsetChangedDelegate(RaptureGearsetModule* module, uint gearsetId, byte glamourPlateId);
+    public delegate int GearsetChangedDelegate(RaptureGearsetModule* module, int gearsetId, byte glamourPlateId);
     internal static Hook<GearsetChangedDelegate> ProcessGearsetChangeHook = null!;
 
-    private nint GearsetChangedDetour(RaptureGearsetModule* module, uint gearsetId, byte glamourPlateId)
+    private int GearsetChangedDetour(RaptureGearsetModule* module, int gearsetId, byte glamourPlateId)
     {
         // Store previous, then perform the original to process the change.
         var prevGearsetIdx = module->CurrentGearsetIndex;
         var prevJob = module->GetGearset(prevGearsetIdx)->ClassJob;
         var ret = ProcessGearsetChangeHook.Original(module, gearsetId, glamourPlateId);
         // Then get the set gearsetIdx
-        var newGearsetEntry = module->GetGearset((int)gearsetId);
+        var newGearsetEntry = module->GetGearset(gearsetId);
         var newJobId = newGearsetEntry->ClassJob;
 
-        _mediator.Publish(new GearsetChangedMessage(prevGearsetIdx, prevJob, (int)gearsetId, newJobId));
+        _mediator.Publish(new GearsetChangedMessage(prevGearsetIdx, prevJob, gearsetId, newJobId));
         return ret;
     }
 }
